+++ frontend_nustly/src/features/Product/CreateProduct/index.ts
@@ -1,0 +1,1 @@
+export { default as  CreateProduct } from './ui/CreateProduct.vue';
