+++ frontend_nustly/src/features/Product/ShowProduct/ui/ShowProduct.vue
@@ -1,0 +1,13 @@
+<template>
+  <div>{{ product }}</div>
+</template>
+
+<script setup lang="ts">
+import { useProductsStore } from '@/shered/store/products';
+import { storeToRefs } from 'pinia';
+
+const productsStore = useProductsStore();
+const { product } = storeToRefs(productsStore);
+</script>
+
+<style src="./ShowProduct.style.scss" lang="scss" scoped></style>
