--- frontend_nustly/src/features/Product/ui/index.ts
@@ -1,1 +1,0 @@
-export { default as ShowProduct } from './ShowProduct/ShowProduct.vue'
