+++ frontend_nustly/src/pages/ProductsPage/ProductsPage.vue
@@ -23,7 +23,11 @@
       v-if="products.length"
     >
       <template v-for="elem in products" :key="elem.id">
-        <ProductItem :elemProduct="elem"/>
+        <ProductItem 
+          :elemProduct="elem"
+          :handlerShowProductDialogVisible="handlerShowProductDialogVisible"
+          :getProduct="getProduct"
+        />
       </template>
     </div>
     <div 
@@ -43,48 +47,16 @@
     <h2 v-if="!isLoading && (products.length === 0)" class="products__error">
       Данных нет
     </h2>
-    <dialog-window v-model:show="creatProductDialogVisible">
-      <form @submit.prevent :style="{padding: '10px'}">
-        <h1 :style="{marginBottom: '30px'}">Создать новый товар</h1>
-        <input-elem    
-          v-model="productInfo.title"
-          :typeInput="'text'"
-          :placeholderInput="'Название'"
-        />
-        <text-area 
-          v-model="productInfo.description" 
-          :placeholderInput="'Описание'"
-        />
-        <input-elem    
-          v-model="productInfo.manufacturer"
-          :typeInput="'text'"
-          :placeholderInput="'Производитель'"
-        />
-        <input-elem    
-          v-model="productInfo.price"
-          :typeInput="'number'"
-          :placeholderInput="'Цена'"
-        />
-        <input 
-          type="file" 
-          @change="onFileChange" 
-          :style="{margin: '15px 0'}"
-        >
-        <button-elem
-          :clName="null"
-          :title="'Отправить'"
-          :handler="submit"
-          :width="'100%'"
-          :height="'48px'"
-          :background="'#70C05B'"
-          :textColor="null"
-          :fontSize="null"
-          :fontWeight="null"
-          :margin="'20px 0 0 0'"
-          :borderRadius="'10px'"
-          :icon="null"
-        />
-      </form>
+    <dialog-window v-model:show="createProductDialogVisible">
+      <CreateProduct
+        :categoryId="categoryId"
+        :postCreateProduct="postCreateProduct"
+        :getProducts="getProducts"
+        :handlerCreateProductDialogVisible="handlerCreateProductDialogVisible"
+      />
+    </dialog-window>
+    <dialog-window v-model:show="showProductDialogVisible">
+      <ShowProduct/>
     </dialog-window>
   </div>
 </template>
@@ -97,14 +69,15 @@ import { storeToRefs } from 'pinia';
 import { computed, onMounted, reactive, ref } from 'vue';
 import ProductItem from '@/entities/productItem/ProductItem.vue';
 import { useRoute } from 'vue-router';
+import { CreateProduct } from '@/features/Product/CreateProduct';
+import { ShowProduct } from '@/features/Product/ShowProduct';
 
 const route = useRoute();
 const categoryId = route.params.products_id;
 
-
 //catalog store
 const productsStore = useProductsStore();
-const { isLoading, errorMessage, products, product, totalPages, page  } = storeToRefs(productsStore);
+const { isLoading, errorMessage, products, totalPages, page  } = storeToRefs(productsStore);
 const { getProducts, getProduct, postCreateProduct } = productsStore;
 
 //theme store
@@ -120,31 +93,11 @@ const authStore = useAuthStore();
 const { role } = storeToRefs(authStore);
 
 //dialog window
-const creatProductDialogVisible = ref(false);
+const createProductDialogVisible = ref(false);
 const showProductDialogVisible = ref(false);
 
-const productInfo = reactive({
-  id_categories: categoryId,
-  title: '',
-  description: '',
-  manufacturer: '',
-  price: null,
-});
-const fileImg = ref(null);
-
-
-const productForm = computed(() => {
-  return {
-    id_categories: categoryId,
-    title: productInfo.title,
-    description: productInfo.description,
-    manufacturer: productInfo.manufacturer,
-    price: productInfo.price
-  };
-});
-
 const handlerCreateProductDialogVisible = () => {
-  creatProductDialogVisible.value = !creatProductDialogVisible.value;
+  createProductDialogVisible.value = !createProductDialogVisible.value;
 };
 
 const handlerShowProductDialogVisible = () => {
@@ -154,25 +107,6 @@ const handlerShowProductDialogVisible = () => {
 //methods
 const changePage = (currentPage: number) => {
   page.value = currentPage;
-};
-
-const onFileChange = (event: any) => {
-  fileImg.value = event.target.files[0];
-};
-
-const submit = async () => {
-  if (!fileImg.value) {
-    alert('Фотография не прекреплена');
-  } else {
-    await postCreateProduct(productForm.value, fileImg.value);
-    getProducts(categoryId);
-    handlerCreateProductDialogVisible();
-    fileImg.value = null;
-    productInfo.title = '';
-    productInfo.description = '';
-    productInfo.manufacturer = '';
-    productInfo.price = null;
-  }
 };
 
 onMounted(() => {
